requirements.txt
Django==3.1
psycopg2-binary>=2.8
gunicorn
channelsmultiplexer==0.0.2
djangochannelsrestframework==0.1.0
djangorestframework==3.11.1